+++ lib/fluent/dto/logEventsJson.rb
@@ -1,4 +1,4 @@
-## Copyright (c) 2021, 2022  Oracle and/or its affiliates.
+## Copyright (c) 2021, 2024  Oracle and/or its affiliates.
 ## The Universal Permissive License (UPL), Version 1.0 as shown at https://oss.oracle.com/licenses/upl/
 
 require_relative './logEvents'
@@ -20,5 +20,5 @@ class LogEventsJson
       end
     }.compact
   end
-  
+
 end