+++ lib/fluent/metrics/metricsLabels.rb
@@ -1,5 +1,8 @@
+## Copyright (c) 2021, 2024  Oracle and/or its affiliates.
+## The Universal Permissive License (UPL), Version 1.0 as shown at https://oss.oracle.com/licenses/upl/
+
 class MetricsLabels
-  attr_accessor :worker_id, :tag, :logGroupId, :logSourceName, :logSet, :invalid_reason, :records_valid, :records_per_tag, :latency
+  attr_accessor :worker_id, :tag, :logGroupId, :logSourceName, :logSet, :invalid_reason, :records_valid, :records_per_tag, :latency,:timezone
   def initialize
       @worker_id = nil
       @tag = nil
@@ -10,5 +13,6 @@ class MetricsLabels
       @records_valid = 0
       @records_per_tag = 0
       @latency = 0
+      @timezone = nil
     end
 end